// Web/RMA-Web-Shell/projects/medicare/src/app/preauth-manager/models/prosthetic-document-states.ts
import { ProstheticQuotationTypeEnum } from "../../medi-manager/enums/prosthetic-quotation-type-enum";

export class ProstheticDocumentStates {
    ProstheticOrthoticDocsOver2YrsNonPension: boolean;
    OrthoticDocsOver2YrsNonPension: boolean;
    ProstheticDocsOver2YrsPension: boolean;
    selectedQuotationType:ProstheticQuotationTypeEnum
}